backend/requirements.txt
fastapi>=0.116.1,<0.117
uvicorn[standard]>=0.35.0,<0.36
sqlalchemy>=2.0.43,<3
pymysql>=1.1.1,<2
pydantic-settings>=2.10.1,<3
numpy>=2.2,<3
pandas>=2.2,<4
scikit-learn>=1.7,<1.9
tensorflow-cpu==2.17.0; python_version < "3.13"
python-dateutil>=2.9.0.post0,<3
joblib>=1.5.2,<2
python-multipart>=0.0.20,<0.1
cryptography>=45,<46
PyJWT>=2.10.1,<3
